Separate the recycler setup into a method in MainActivity class
