+++ app/src/main/AndroidManifest.xml
@@ -11,11 +11,11 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".views.WebDetailActivity"></activity>
+        <activity android:name=".views.WebDetailActivity" />
         <activity
             android:name=".views.PokemonDetailedActivity"
-            android:parentActivityName=".MainActivity"></activity>
-        <activity android:name=".MainActivity">
+            android:parentActivityName=".views.MainActivity"></activity>
+        <activity android:name=".views.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
